[test-upstream] Move _NoDefault to pdcompat; fix doc build
