+++ xarray/core/pdcompat.py
@@ -35,6 +35,9 @@
 # OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 from __future__ import annotations
 
+from enum import Enum
+from typing import Literal
+
 
 def count_not_none(*args) -> int:
     """Compute the number of non-None arguments.
@@ -42,3 +45,26 @@ def count_not_none(*args) -> int:
     Copied from pandas.core.common.count_not_none (not part of the public API)
     """
     return sum(arg is not None for arg in args)
+
+
+class _NoDefault(Enum):
+    """Used by pandas to specify a default value for a deprecated argument.
+    Copied from pandas._libs.lib._NoDefault.
+
+    See also:
+    - pandas-dev/pandas#30788
+    - pandas-dev/pandas#40684
+    - pandas-dev/pandas#40715
+    - pandas-dev/pandas#47045
+    """
+
+    no_default = "NO_DEFAULT"
+
+    def __repr__(self) -> str:
+        return "<no_default>"
+
+
+no_default = (
+    _NoDefault.no_default
+)  # Sentinel indicating the default value following pandas
+NoDefault = Literal[_NoDefault.no_default]  # For typing following pandas
